share parameterization code between n plus one api and consecutive http
