+++ src/sentry/utils/performance_issues/detectors/consecutive_http_detector.py
@@ -1,9 +1,6 @@
 from __future__ import annotations
 
-import hashlib
-import re
 from datetime import timedelta
-from urllib.parse import parse_qs, urlparse
 
 from sentry import features
 from sentry.issues.grouptype import PerformanceConsecutiveHTTPQueriesGroupType
@@ -12,33 +9,12 @@ from sentry.models import Organization, Project
 from ..base import (
     DetectorType,
     PerformanceDetector,
+    fingerprint_http_spans,
     get_duration_between_spans,
     get_span_duration,
-    get_url_from_span,
 )
 from ..performance_problem import PerformanceProblem
 from ..types import Span
-
-URL_PARAMETER_REGEX = re.compile(
-    r"""(?x)
-    (?P<uuid>
-        \b
-            [0-9a-fA-F]{8}-
-            [0-9a-fA-F]{4}-
-            [0-9a-fA-F]{4}-
-            [0-9a-fA-F]{4}-
-            [0-9a-fA-F]{12}
-        \b
-    ) |
-    (?P<hashlike>
-        \b[0-9a-fA-F]{10}([0-9a-fA-F]{14})?([0-9a-fA-F]{8})?([0-9a-fA-F]{8})?\b
-    ) |
-    (?P<int>
-        -\d+\b |
-        \b\d+\b
-    )
-"""
-)  # Adapted from message.py
 
 
 class ConsecutiveHTTPSpanDetector(PerformanceDetector):
@@ -155,16 +131,7 @@ class ConsecutiveHTTPSpanDetector(PerformanceDetector):
         return True
 
     def _fingerprint(self) -> str:
-        url_paths = []
-        for http_span in self.consecutive_http_spans:
-            url = get_url_from_span(http_span)
-            parametrized_url = self.parameterize_url(url)
-            path = urlparse(parametrized_url).path
-            if path not in url_paths:
-                url_paths.append(path)
-        url_paths.sort()
-
-        hashed_url_paths = hashlib.sha1(("-".join(url_paths)).encode("utf8")).hexdigest()
+        hashed_url_paths = fingerprint_http_spans(self.consecutive_http_spans)
         return f"1-{PerformanceConsecutiveHTTPQueriesGroupType.type_id}-{hashed_url_paths}"
 
     def on_complete(self) -> None:
@@ -174,38 +141,6 @@ class ConsecutiveHTTPSpanDetector(PerformanceDetector):
         return features.has(
             "organizations:performance-consecutive-http-detector", organization, actor=None
         )
-
-    @staticmethod
-    def parameterize_url(url: str) -> str:
-        parsed_url = urlparse(str(url))
-
-        protocol_fragments = []
-        if parsed_url.scheme:
-            protocol_fragments.append(parsed_url.scheme)
-            protocol_fragments.append("://")
-
-        host_fragments = []
-        for fragment in parsed_url.netloc.split("."):
-            host_fragments.append(str(fragment))
-
-        path_fragments = []
-        for fragment in parsed_url.path.split("/"):
-            if URL_PARAMETER_REGEX.search(fragment):
-                path_fragments.append("*")
-            else:
-                path_fragments.append(str(fragment))
-
-        query = parse_qs(parsed_url.query)
-
-        return "".join(
-            [
-                "".join(protocol_fragments),
-                ".".join(host_fragments),
-                "/".join(path_fragments),
-                "?",
-                "&".join(sorted([f"{key}=*" for key in query.keys()])),
-            ]
-        ).rstrip("?")
 
     def is_creation_allowed_for_project(self, project: Project) -> bool:
         return True
